backend/requirements.txt
fastapi
uvicorn
python-multipart
markdown
python-jose[cryptography]
passlib[bcrypt]